[refactor] Create todo module
- app.js에 있는 todo 실행 로직을 모듈로 분리
